Move test_write_forward_msg_reraises_websocket_closed_error to BrowserWebSocketHandlerTest
