+++ lib/tests/streamlit/web/server/server_test.py
@@ -46,10 +46,6 @@ from streamlit.web.server.server import RetriesExceeded
 from streamlit.web.server.server import Server
 from streamlit.web.server.server import State
 from streamlit.web.server.server import StaticFileHandler
-from streamlit.web.server.server import (
-    BrowserWebSocketHandler,
-    SessionClientDisconnectedError,
-)
 from streamlit.web.server.server import start_listening
 from streamlit.web.server.server_util import (
     is_cacheable_msg,
@@ -418,40 +414,6 @@ class ServerTest(ServerTestCase):
                 self.assertIsNone(
                     self.server._get_session_info(session_info.session.id)
                 )
-
-    @tornado.testing.gen_test
-    async def test_write_forward_msg_reraises_websocket_closed_error(self):
-        """`write_forward_msg` should re-raise WebSocketClosedError as
-        as SessionClientDisconnectedError.
-        """
-
-        with self._patch_app_session():
-            await self.start_server_loop()
-            await self.ws_connect()
-
-            # Get our connected BrowserWebSocketHandler
-            session_info = list(self.server._session_info_by_id.values())[0]
-            websocket_handler = session_info.client
-            self.assertIsInstance(websocket_handler, BrowserWebSocketHandler)
-
-            # Patch _BrowserWebSocketHandler.write_message to raise an error
-            with mock.patch.object(
-                websocket_handler, "write_message"
-            ) as write_message_mock:
-                write_message_mock.side_effect = tornado.websocket.WebSocketClosedError
-
-                msg = ForwardMsg()
-                msg.script_finished = (
-                    ForwardMsg.ScriptFinishedStatus.FINISHED_SUCCESSFULLY
-                )
-
-                # Send a ForwardMsg. write_message will raise a
-                # WebSocketClosedError, and write_forward_msg should re-raise
-                # it as a SessionClientDisconnectedError.
-                with self.assertRaises(SessionClientDisconnectedError):
-                    websocket_handler.write_forward_msg(msg)
-
-                write_message_mock.assert_called_once()
 
     @tornado.testing.gen_test
     async def test_is_active_session(self):
